+++ lib/img_builder.rb
@@ -84,8 +84,8 @@ class ImgBuilder
   end
 
   def generate_wrapper
-    classes = ("imgWrapper #{@props.img_display} #{@props.align} #{@props.attr_size_class} " +
-              "#{@props.attr_max_width_class} #{@props.wrapper_class}").squish
+    classes = ("imgWrapper #{@props.img_display} #{@props.align} #{@props.attr_size_class} " \
+                "#{@props.attr_max_width_class} #{@props.wrapper_class}").squish
     styles = "#{@props.attr_width_style} #{@props.attr_max_width_style} #{@props.wrapper_style}".squish
     <<~END_HTML.remove_blank_lines
       <div class='#{classes}' style='#{styles}'>