+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   This requires new CSS classes to be defined:
   `max_eighthsize`, `max_fullsize`, `max_halfsize`, and `max_quartersize`.
 * More CSS classes defined and more tests added to the Demo app to support `max-width`.
+* `width='calc(something)'` added.
+
 
 ## 0.2.8 / 2025-05-16